update: Inbox update

+++ src/Host/NB12.Boilerplate.Host.Shared/Ops/Dtos/OpsRetentionStatusDto.cs
@@ -3,7 +3,7 @@
     public sealed record OpsRetentionStatusDto(
         bool Enabled,
         DateTime? LastRunAtUtc,
-        int? LastDeletedAuditLogs,
-        int? LastDeletedErrorLogs,
+        long? LastDeletedAuditLogs,
+        long? LastDeletedErrorLogs,
         string? LastError);
 }
