+++ src/Modules/Audit/NB12.Boilerplate.Modules.Audit.Application/Responses/AuditRetentionStatusDto.cs
@@ -3,7 +3,7 @@
     public sealed record AuditRetentionStatusDto(
         bool Enabled,
         DateTime? LastRunAtUtc,
-        int? LastDeletedAuditLogs,
-        int? LastDeletedErrorLogs,
+        long? LastDeletedAuditLogs,
+        long? LastDeletedErrorLogs,
         string? LastError);
 }
